introduction view model and connected to introduction screen
