+++ app/src/main/java/com/example/hydrosaurus/viewModels/FirestoreViewModel.kt
@@ -7,29 +7,12 @@ import com.google.firebase.FirebaseApp
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
-class FirestoreViewModel() : ViewModel() {
+open class FirestoreViewModel() : ViewModel() {
 
     fun setup(context: Context) {
         FirebaseApp.initializeApp(context)
     }
 
-    fun createUserDocument(name: String, email: String) {
-        val auth = FirebaseAuth.getInstance()
-        val uid = auth.currentUser?.uid
-        val db = FirebaseFirestore.getInstance()
-
-        if (uid != null) {
-            val user = hashMapOf(
-                "name" to name,
-                "email" to email,
-                "totalWater" to 0
-            )
-            db.collection("users").document(uid).set(user)
-                .addOnSuccessListener { Log.d("Firestore", "User data saved successfully") }
-                .addOnFailureListener { e -> Log.w("Firestore", "Error saving user data", e) }
-
-        }
-    }
 
     fun readUserDocument(){
         val auth = FirebaseAuth.getInstance()
